Move all the templates into directory
