+++ packages/zed/taiga.json
@@ -4,17 +4,119 @@
   "name": "Taiga",
   "themes": [
     {
-      "name": "Tagia - dark",
+      "name": "Taiga - dark",
       "appearance": "dark",
       "style": {
-        "background": "#000924"
+        "background": "#000924",
+        "border": "#000924",
+        "border.focused": "#283b5a",
+        "border.selected": "#00782D",
+        "border.disabled": "#0F171F",
+        "border.transparent": "#10213e",
+        "editor.background": "#000924",
+        "editor.foreground": "#D4DFEB",
+        "editor.gutter.background": "#000924",
+        "editor.line_number": "#D4DFEBBF",
+        "editor.active_line_number": "#93F8B8",
+        "editor.invisible": "#D4DFEB80",
+        "status_bar.background": "#10213e",
+        "tab_bar.background": "#000924",
+        "tab.active_background": "#10213e",
+        "tab.inactive_background": "#000924",
+        "toolbar.background": "#10213e",
+        "border.variant": "#10213e",
+        "title_bar.background": "#000924",
+        "link_text.hover": "#00782D",
+        "panel.background": "#000924",
+        "elevated_surface.background": "#283b5a",
+        "drop_target.background": "#10213e80",
+        "element.hover": "#10213e80",
+        "element.selected": "#283b5a80",
+        "conflict": "#945CE1",
+        "conflict.border": "#945CE1",
+        "conflict.background": "#945CE180",
+        "editor.active_line.background": "#283b5a80",
+        "modified": "#43B2AC",
+        "modified.background": "#43B2AC",
+        "modified.border": "#43B2AC",
+        "deleted": "#FF6551",
+        "deleted.background": "#BA0D01",
+        "deleted.border": "#BA0D01",
+        "warning": "#CD9200",
+        "warning.background": "#825B00",
+        "warning.border": "#825B00",
+        "created": "#51B67A",
+        "created.background": "#007742",
+        "created.border": "#007742",
+        "text": "#D4DFEB",
+        "text.placeholder": "#D4DFEB80",
+        "text.disabled": "#D4DFEB80",
+        "text.accent": "#51B67A",
+        "text.muted": "#D4DFEBBF",
+        "ghost_element.hover": "#42567780",
+        "ghost_element.active": "#425677",
+        "ghost_element.selected": "#425677",
+        "ghost_element.disabled": "#D4DFEB40",
+        "element.background": "#BA0D01",
+        "element.active": ""
       }
     },
     {
-      "name": "Tagia - light",
+      "name": "Taiga - light",
       "appearance": "light",
       "style": {
-        "background": "#E5EFFF"
+        "background": "#E5EFFF",
+        "border": "#E5EFFF",
+        "border.focused": "#7A9FDD",
+        "border.selected": "#00782D",
+        "border.disabled": "#404952",
+        "border.transparent": "#98BFFE",
+        "editor.background": "#E5EFFF",
+        "editor.foreground": "#0F171F",
+        "editor.gutter.background": "#E5EFFF",
+        "editor.line_number": "#0F171FBF",
+        "editor.active_line_number": "#00381C",
+        "editor.invisible": "#0F171F80",
+        "status_bar.background": "#98BFFE",
+        "tab_bar.background": "#E5EFFF",
+        "tab.active_background": "#98BFFE",
+        "tab.inactive_background": "#E5EFFF",
+        "toolbar.background": "#98BFFE",
+        "border.variant": "#98BFFE",
+        "title_bar.background": "#E5EFFF",
+        "link_text.hover": "#00782D",
+        "panel.background": "#E5EFFF",
+        "elevated_surface.background": "#7A9FDD",
+        "drop_target.background": "#98BFFE80",
+        "element.hover": "#98BFFE80",
+        "element.selected": "#7A9FDD80",
+        "conflict": "#945CE1",
+        "conflict.border": "#945CE1",
+        "conflict.background": "#945CE180",
+        "editor.active_line.background": "#7A9FDD80",
+        "modified": "#00736E",
+        "modified.background": "#00736E",
+        "modified.border": "#00736E",
+        "deleted": "#BA0D01",
+        "deleted.background": "#FF6551",
+        "deleted.border": "#FF6551",
+        "warning": "#825B00",
+        "warning.background": "#CD9200",
+        "warning.border": "#CD9200",
+        "created": "#007742",
+        "created.background": "#51B67A",
+        "created.border": "#51B67A",
+        "text": "#0F171F",
+        "text.placeholder": "#0F171F80",
+        "text.disabled": "#0F171F80",
+        "text.accent": "#007742",
+        "text.muted": "#0F171FBF",
+        "ghost_element.hover": "#5C80BC80",
+        "ghost_element.active": "#5C80BC",
+        "ghost_element.selected": "#5C80BC",
+        "ghost_element.disabled": "#0F171F40",
+        "element.background": "#FF6551",
+        "element.active": ""
       }
     }
   ]
