+++ src/main.ts
@@ -1,8 +1,8 @@
 import { resolve } from "node:path";
-import { generate as vscode } from "./vscode/build.ts";
-import { generate as ghostty } from "./ghostty/build.ts";
-import { generate as warp } from "./warp/build.ts";
-import { generate as zed } from "./zed/build.ts";
+import { generate as vscode } from "./templates/vscode/build.ts";
+import { generate as ghostty } from "./templates/ghostty/build.ts";
+import { generate as warp } from "./templates/warp/build.ts";
+import { generate as zed } from "./templates/zed/build.ts";
 import darkColours from "./colours/dark-colours.json" with { type: "json" };
 import lightColours from "./colours/light-colours.json" with { type: "json" };
