+++ src/templates/ghostty/build.ts
