+++ src/templates/ghostty/theme.ini
