+++ src/templates/vscode/color-theme.json
