+++ src/templates/zed/build.ts
@@ -2,8 +2,8 @@ import { writeFile } from "node:fs";
 import { resolve } from "node:path";
 import mustache from "mustache";
 
-import { opacity } from "../utils/opacity.ts";
-import { darkLight } from "../utils/darkLight.ts";
+import { opacity } from "../../utils/opacity.ts";
+import { darkLight } from "../../utils/darkLight.ts";
 import zedTheme from "../zed/theme.json" with { type: "json" };
 
 /**
