+++ src/templates/zed/theme.json
@@ -1,0 +1,65 @@
+{
+	"$schema": "https://zed.dev/schema/themes/v0.2.0.json",
+	"author": "Josh Dales",
+	"name": "Taiga",
+	"themes": [
+		{
+			"name": "Taiga - {{ type }}",
+			"appearance": "{{type}}",
+			"style": {
+				"background": "{{ ui.0 }}",
+				"border": "{{ui.0}}",
+				"border.focused": "{{ui.2}}",
+				"border.selected": "{{ui.highlight}}",
+				"border.disabled": "{{ui.shadow}}",
+				"border.transparent": "{{ui.1}}",
+				"editor.background": "{{ ui.0 }}",
+				"editor.foreground": "{{ ui.text }}",
+				"editor.gutter.background": "{{ ui.0 }}",
+				"editor.line_number": "{{ui.text}}{{#opacity}}75{{/opacity}}",
+				"editor.active_line_number": "{{green.4}}",
+				"editor.invisible": "{{ ui.text }}{{#opacity}}50{{/opacity}}",
+				"status_bar.background": "{{ui.1}}",
+				"tab_bar.background": "{{ui.0}}",
+				"tab.active_background": "{{ui.1}}",
+				"tab.inactive_background": "{{ui.0}}",
+				"toolbar.background": "{{ui.1}}",
+				"border.variant": "{{ui.1}}",
+				"title_bar.background": "{{ui.0}}",
+				"link_text.hover": "{{ui.highlight}}",
+				"panel.background": "{{ui.0}}",
+				"elevated_surface.background": "{{ui.2}}",
+				"drop_target.background": "{{ui.1}}{{#opacity}}50{{/opacity}}",
+				"element.hover": "{{ui.1}}{{#opacity}}50{{/opacity}}",
+				"element.selected": "{{ui.2}}{{#opacity}}50{{/opacity}}",
+				"conflict": "{{purple.1}}",
+				"conflict.border": "{{purple.1}}",
+				"conflict.background": "{{purple.1}}{{#opacity}}50{{/opacity}}",
+				"editor.active_line.background": "{{ui.2}}{{#opacity}}50{{/opacity}}",
+				"modified": "{{teal.2}}",
+				"modified.background": "{{teal.2}}",
+				"modified.border": "{{teal.2}}",
+				"deleted": "{{red.2}}",
+				"deleted.background": "{{red.0}}",
+				"deleted.border": "{{red.0}}",
+				"warning": "{{yellow.2}}",
+				"warning.background": "{{yellow.0}}",
+				"warning.border": "{{yellow.0}}",
+				"created": "{{green.2}}",
+				"created.background": "{{green.0}}",
+				"created.border": "{{green.0}}",
+				"text": "{{ui.text}}",
+				"text.placeholder": "{{ui.text}}{{#opacity}}50{{/opacity}}",
+				"text.disabled": "{{ui.text}}{{#opacity}}50{{/opacity}}",
+				"text.accent": "{{green.2}}",
+				"text.muted": "{{ui.text}}{{#opacity}}75{{/opacity}}",
+				"ghost_element.hover": "{{ui.3}}{{#opacity}}50{{/opacity}}",
+				"ghost_element.active": "{{ui.3}}",
+				"ghost_element.selected": "{{ui.3}}",
+				"ghost_element.disabled": "{{ui.text}}{{#opacity}}25{{/opacity}}",
+				"element.background": "{{red.0}}",
+				"element.active": "{{ui.red}}"
+			}
+		}
+	]
+}
