--- src/zed/theme.json
@@ -1,15 +1,0 @@
-{
-	"$schema": "https://zed.dev/schema/themes/v0.2.0.json",
-	"author": "Josh Dales",
-	"name": "Taiga",
-	"themes": [
-		{
-			"name": "Tagia - {{ type }}",
-			"appearance": "{{type}}",
-			"style": {
-				"background": "{{ ui.0 }}",
-				"editor.background": "{{ ui.0 }}"
-			}
-		}
-	]
-}
